Move types to proper files
