+++ packages/twenty-server/src/modules/workflow/common/types/workflow-settings.type.ts
@@ -1,0 +1,14 @@
+export type WorkflowCodeSettingsType = {
+  serverlessFunctionId: string;
+};
+
+export type WorkflowSettingsType = {
+  errorHandlingOptions: {
+    retryOnFailure: {
+      value: boolean;
+    };
+    continueOnFailure: {
+      value: boolean;
+    };
+  };
+} & WorkflowCodeSettingsType;
