+++ packages/twenty-server/src/modules/workflow/workflow-trigger/workflow-trigger.service.ts
@@ -2,16 +2,16 @@ import { Injectable } from '@nestjs/common';
 
 import { TwentyORMGlobalManager } from 'src/engine/twenty-orm/twenty-orm-global.manager';
 import { WorkflowEventListenerWorkspaceEntity } from 'src/modules/workflow/common/standard-objects/workflow-event-listener.workspace-entity';
-import {
-  WorkflowDatabaseEventTrigger,
-  WorkflowTrigger,
-  WorkflowTriggerType,
-  WorkflowVersionWorkspaceEntity,
-} from 'src/modules/workflow/common/standard-objects/workflow-version.workspace-entity';
+import { WorkflowVersionWorkspaceEntity } from 'src/modules/workflow/common/standard-objects/workflow-version.workspace-entity';
 import {
   WorkflowTriggerException,
   WorkflowTriggerExceptionCode,
 } from 'src/modules/workflow/workflow-trigger/workflow-trigger.exception';
+import {
+  WorkflowDatabaseEventTrigger,
+  WorkflowTrigger,
+  WorkflowTriggerType,
+} from 'src/modules/workflow/common/types/workflow-trigger.type';
 
 @Injectable()
 export class WorkflowTriggerService {
